requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==2.3.7
requests==2.31.0
Pillow==10.1.0

# Database dependencies
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Security dependencies  
cryptography==41.0.7
python-dotenv==1.0.0

# Production dependencies
gunicorn==21.2.0

# Development dependencies
python-json-logger==2.0.7